systemProject/geminiOCR/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.5

# PDF processing
pdf2image==1.17.0
Pillow==11.1.0

# Google Gemini
google-generativeai==0.8.4

# Environment
python-dotenv==1.0.1
